+++ src/rle.rs
@@ -265,7 +265,7 @@ impl RunLengthEncoding {
         }
     }
 
-    pub fn collect_indices(&self) -> Box<[usize]> {
+    pub fn collect_indices(&self) -> Box<[u32]> {
         // Create the blocks iterator
         let mut blocks_iter = self
             .blocks
@@ -273,20 +273,20 @@ impl RunLengthEncoding {
             .map(|block_u16| Block::from_u16(*block_u16));
 
         // Initialize curr_i and the return value
-        let mut curr_i = 0;
+        let mut curr_i = 0_u32;
         let mut indices = vec![];
 
         while let Some(block) = blocks_iter.next() {
             match block {
-                Block::Zeros(zeroes_count) => curr_i += zeroes_count as usize,
+                Block::Zeros(zeroes_count) => curr_i += zeroes_count as u32,
                 Block::Ones(ones_count) => {
-                    let ones_count = ones_count as usize;
+                    let ones_count = ones_count as u32;
                     indices.extend(curr_i..curr_i + ones_count);
                     curr_i += ones_count;
                 }
                 Block::Uncompressed(bits) => {
-                    indices.extend(BitIter::from(bits).map(|i| i + curr_i));
-                    curr_i += MAX_UNCOMPRESSED_BITS;
+                    indices.extend(BitIter::from(bits).map(|i| i as u32 + curr_i));
+                    curr_i += MAX_UNCOMPRESSED_BITS as u32;
                 }
             }
         }
